config/settings.py
import os 
from dotenv import load_dotenv


load_dotenv()

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
TAVILY_API_KEY = os.getenv("TAVILY_API_KEY")
MONGO_API_BASE = "http://192.168.1.64:5000/api/v1/chats"